Converted np.float to np.float64
`np.float` is a deprecated command that yields the following error:

```
DeprecationWarning: np.float is a deprecated alias for the builtin float. To silence this warning, use float by itself. Doing this will not modify any behavior and is safe. If you specifically wanted the numpy scalar type, use np.float64 here.
```

I changed it to `np.float64` and the evaluation script started working
again. Thank you so much for putting this repository together!